requirements.txt
transformers
torch
qdrant-client
pillow
requests